Structure update (#25)
* add /en/ as a redirect to /

* remove the <base> element, relink resources in code and font page based on that

* remove full links to section headers
This caused a small issue in the German course, but now that we aren't using <base>, I can easily make them just section links

* make /en/ redirect with status code 301 Permanently Moved

* forgot that it's not TS

* remove duplicate sound entries

* have PrevNext go to the parent page

* remove base url from config

* fix links to other courses on translation page

* change // ... comment to /* ... */

* change language codes to two periods on course pages (for all courses)
This is for convenience of translating the courses, so links to other course pages no longer need to provide a language code. The only page where that needs to be done is the index (since those pages are based on the English index, which is at / instead of /en/). Though indexes in other languages can still work if "/ru" or "/yue" is replaced with ".."

* relink course pages in indexes
This makes links to course pages in other languages work properly.

* convert relative links to absolute links
This was requested by kala Asi: "[using relative links] forces the markdown writer to be aware of the peculiarities of index.html file generation"

* add trailing comma to config item